+++ gitime/database.py
@@ -22,14 +22,14 @@ DB_NAME = PATHCHAR.join((DB_DIR, 'gitime.db'))
 
 
 def set_unix_permissions(path):
-    """ Linux and Mac users will probably install using sudo,
+    """ Linux and Mac users might install using sudo,
         so the user needs to be granted read and write access
         to the database
     """
     import pwd
     uname = os.getenv("SUDO_USER") or os.getenv("USER")
-    os.chown(DB_DIR, pwd.getpwnam(uname).pw_uid, pwd.getpwnam(uname).pw_gid)
-    os.chmod(DB_DIR, stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
+    os.chown(path, pwd.getpwnam(uname).pw_uid, pwd.getpwnam(uname).pw_gid)
+    os.chmod(path, stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
 
 
 def db_exists():